+++ src/Divider/Divider.tsx
@@ -1,8 +1,9 @@
 import React, { FC, memo } from 'react'
 import styled, { css } from 'styled-components'
-import { MarginProps } from '../utils/space'
+import { TransientProps } from 'utils/utilTypes'
 import { Box } from '../Box'
 import { Color, theme } from '../theme'
+import { MarginProps } from '../utils/space'
 
 export type DividerProps = {
   maxWidth?: string
@@ -17,19 +18,25 @@ export const Divider: FC<DividerProps> = memo(function Divider({
   ...marginProps
 }) {
   return (
-    <Line maxWidth={maxWidth} height={height} color={color} {...marginProps} />
+    <Line
+      $maxWidth={maxWidth}
+      $height={height}
+      $color={color}
+      {...marginProps}
+    />
   )
 })
 
-const Line = styled(Box)<{ color: Color }>`
-  ${({ color }) =>
-    color
-      ? css`
-          background: ${theme.colors[color]};
-        `
-      : ''}
+const Line = styled(Box)<
+  TransientProps<Pick<DividerProps, 'color' | 'height' | 'maxWidth'>>
+>`
+  ${({ $color }) =>
+    $color &&
+    css`
+      background: ${theme.colors[$color]};
+    `}
   display: block;
-  height: ${({ height }) => height};
-  max-width: ${({ maxWidth }) => maxWidth};
+  height: ${({ $height }) => $height};
+  max-width: ${({ $maxWidth }) => $maxWidth};
   width: 100%;
 `