+++ src/Dropdown/Dropdown.tsx
@@ -105,7 +105,7 @@ export const Dropdown = forwardRef(function Dropdown(
       <Box flex alignItems="center">
         {frontIcon && (
           <StyledFrontIcon
-            disabled={disabled}
+            $disabled={disabled}
             render={frontIcon}
             color="sesame"
           />
@@ -120,12 +120,12 @@ export const Dropdown = forwardRef(function Dropdown(
             onInputChange?.(event)
             setValue(value)
           }}
-          error={error}
+          $error={error}
           ref={ref}
           onBlur={onBlur}
           name={name}
-          frontIcon={frontIcon}
-          fallbackStyle={fallbackStyle}
+          $frontIcon={frontIcon}
+          $fallbackStyle={fallbackStyle}
           value={value ? value : ''}
         >
           {hasOptGroups ? (
@@ -178,9 +178,9 @@ const resetSelect = css`
 `
 
 interface SSelect {
-  error: boolean
-  frontIcon?: string
-  fallbackStyle?: boolean
+  $error: boolean
+  $frontIcon?: string
+  $fallbackStyle?: boolean
 }
 
 const StyledSelect = styled.select<SSelect>`
@@ -189,18 +189,18 @@ const StyledSelect = styled.select<SSelect>`
   height: 32px;
 
   cursor: pointer;
-  background-color: ${({ fallbackStyle }) =>
-    fallbackStyle ? theme.colors.custard : theme.colors.cream};
+  background-color: ${({ $fallbackStyle }) =>
+    $fallbackStyle ? theme.colors.custard : theme.colors.cream};
 
   border-radius: 12px;
   padding: 18px 14px;
   border: 2px solid
-    ${({ error }) => (error ? theme.colors.strawberry : theme.colors.oatmeal)};
+    ${({ $error }) => ($error ? theme.colors.strawberry : theme.colors.oatmeal)};
   height: auto;
 
-  ${({ frontIcon }) =>
-    frontIcon &&
-    frontIcon != '' &&
+  ${({ $frontIcon }) =>
+    $frontIcon &&
+    $frontIcon != '' &&
     `
       padding-left: 42px;
     `}
@@ -210,8 +210,8 @@ const StyledSelect = styled.select<SSelect>`
     opacity: 0.5;
   }
 
-  ${({ error }) =>
-    !error &&
+  ${({ $error }) =>
+    !$error &&
     css`
       :not(:disabled) {
         &:hover,