+++ src/Table/components/TableHeader.tsx
@@ -10,29 +10,32 @@ export const TableHeader = <T extends object>({
   headerHeight,
   columnPadding,
   expandable,
+  hasKeyline,
 }: TableHeaderProps<T>) => {
   return (
     <StyledRow>
       {columns.map((column, columnIndex) => (
         <StyledHeaderCell
           key={columnIndex}
-          fixedHeader={fixedHeader}
-          headerHeight={headerHeight}
-          minWidth={column.minWidth}
-          maxWidth={column.maxWidth}
-          headerColor={headerColor}
-          columnPadding={columnPadding}
+          $fixedHeader={fixedHeader}
+          $headerHeight={headerHeight}
+          $minWidth={column.minWidth}
+          $maxWidth={column.maxWidth}
+          $headerColor={headerColor}
+          $columnPadding={columnPadding}
+          $hasKeyline={hasKeyline}
         >
           {column.name}
         </StyledHeaderCell>
       ))}
       {(expandable || rowActions) && (
         <StyledHeaderCell
-          fixedHeader={fixedHeader}
-          stickyCell={true}
-          minWidth={rowActions?.minWidth}
-          headerColor={headerColor}
-          columnPadding={columnPadding}
+          $fixedHeader={fixedHeader}
+          $stickyCell={true}
+          $minWidth={rowActions?.minWidth}
+          $headerColor={headerColor}
+          $columnPadding={columnPadding}
+          $hasKeyline={hasKeyline}
         >
           Actions
         </StyledHeaderCell>